Title: 낮은 지점들, Time: 523ms, Memory: 30MB, Status: Passed - CodeTree

+++ 241005/낮은 지점들/lowest-points.java
@@ -12,7 +12,7 @@ public class Main {
 
         Collection<Integer> values = map.values();
 
-        int ans = 0;
+        long ans = 0L;
         for (int value : values) {
             ans += value;
         }
